requirements.txt
bidict==0.22.0
boto3==1.23.0
botocore==1.26.0
click==8.1.3
DBUtils==3.0.2
dnspython==1.16.0
Flask==2.1.2
Flask-SocketIO==5.1.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
MarkupSafe==2.1.1
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-engineio==4.3.2
python-socketio==5.6.0
s3transfer==0.5.2
six==1.16.0
urllib3==1.26.9
uWSGI==2.0.20
Werkzeug==2.1.2
zope.event==4.5.0
zope.interface==5.4.0
